+++ widgets/SwitchButton.php
@@ -11,21 +11,16 @@ use yii\helpers\Url;
  */
 class SwitchButton extends Widget
 {
-    public $icon;
-    
     public $url;
     
     public function init()
     {
-        // Always use moon icon
-        $this->icon = Icon::get('moon-o');
         $this->url = Url::toRoute('/dark-mode/user/modal/');
     }
     
     public function run()
     {
         return $this->render('switchButton', [
-            'icon' => $this->icon,
             'url' => $this->url,
         ]);
     }