+++ widgets/views/switchButton.php
@@ -1,8 +1,10 @@
 <?php
 
+use humhub\widgets\bootstrap\Button;
+
 ?>
 <div class="btn-group dark-mode">
-    <a id="dark-mode-button" href="#" data-action-click="ui.modal.load" data-action-click-url="<?= $url ?>">
-        <?= $icon ?>
-    </a>
+    <?= Button::asLink()
+        ->icon('moon-o')
+        ->action('ui.modal.load', $url) ?>
 </div>